add seal pipeline
